Crate xng_rs

A Rust wrapper for the XNG API

This crate provides a thin wrapper for the C ABI of FentISS' Xtratum Next Generation (XNG) separation kernel. It allows the implementation of bare metal (no_std) partitions for XNG.

About the Project

This is by no means ready - it is an ongoing progress. While we've already used this together with FentISS' Separation Kernel Emulator (SKE), it was not throughfully tested. While we are engaged with FentISS, there is no official support for this neither from FentISS nor from us. However, if you encounter any problems, please open up an issue. The chances are that we care and try to fix the issue.

Modules

partition
Functions related to the partitionining system

port
Ports allow communication between the partitions in XNG

prelude
The prelude is a collection of all traits and commonly used types in this crate

time
This module adds some notion of time

vcpu
This module contains functions for the virtual CPU

Macros

cstr
Create a NULL terminated string in C representation

to_traceable_error
Convert a xReturnCode_t to an Result<(), XngError>

Structs

XngErrorTrace
An Xng Error with trace information

Enums

XngError
The XNG error type