Module xng_rs::port

Ports allow communication between the partitions in XNG

Types of Ports

There are two different types of ports which differ slightly in semantics.

A Sampling Port retains the last message (if any). It allows for single producer multiple consumer (SPMC). The maximum size of a message is capped at N, shorter messages however are permissible.

In contrast, a Queuing Port retains the last M messages, of which each might be up to N bytes big. The messages are guaranteed to to be served in FIFO order. This type of port is single producer single consumer (SPSC), so only two partitions can use one Queueing Port.

Structs

SamplingPortStatus
The current status of a Sampling Port

SamplingReceiver
Keeps the last (if any) sent value

SamplingSender
Allows to store one message in the port

Type Definitions

SamplingPortId
The type of a sampling ports id