Module xng_rs::prelude

The prelude is a collection of all traits and commonly used types in this crate

For normal use of this crate it is sufficient to glob import only this moduel, e.g. use xng_rs::prelude::*.

Re-exports

pub use cstr_core;

pub use crate::cstr;

pub use crate::partition;

pub use crate::port;

pub use crate::time;

pub use crate::XngError;

Structs

CStr
Representation of a borrowed C string.

Duration
A Duration type to represent a span of time, typically used for system timeouts.